Was passing a copy of an array to ksort() instead of a reference
ksort will only work on data references as it modifies the source data.
We were passing the return of a function which is a copy this caused the
following error:
PHP Strict Standards:  Only variables should be passed by reference in
libraries/template/template.class.php on line 687.

This change takes the copy into a local variable which is sorted and
then used.